chore(tests): move tests in tests folder
